+++ gomock_generator/gomockgenerator/fixtures/int_parameter/int_parameter.go
@@ -1,0 +1,5 @@
+package intparameter
+
+type Service interface {
+	Handle(value int) error
+}